git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C939

+++ inc/Physics/Collider.h
@@ -37,7 +37,7 @@ namespace TikiEngine
 			bool GetTriggerFlag();
 			void SetTriggerFlag(bool triggerFlag);
 
-			void SetCollisionGroup(UInt32 group);
+			void SetCollisionGroup(unsigned short group);
 
 
 			bool GetReady();
